requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.0.5
django-admin-interface==0.14.1
django-colorfield==0.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
djongo==1.3.3
ecdsa==0.14.1
gunicorn==20.0.4
idna==2.10
opentok==2.10.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.1
python-jose==3.2.0
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.2
whitenoise==5.2.0
